+++ src/core/io/event/ResolutionChangedEvent.cpp
@@ -8,7 +8,7 @@ namespace PAX {
     ResolutionChangedEvent::ResolutionChangedEvent(int oldWidth, int oldHeight, int newWidth, int newHeight) :
             _oldWidth(oldWidth), _oldHeight(oldHeight), _newWidth(newWidth), _newHeight(newHeight) {}
 
-    ResolutionChangedEvent::ResolutionChangedEvent(glm::vec2 oldRes, glm::vec2 newRes) :
+    ResolutionChangedEvent::ResolutionChangedEvent(glm::ivec2 oldRes, glm::ivec2 newRes) :
             _oldWidth(oldRes.x), _oldHeight(oldRes.y), _newWidth(newRes.x), _newHeight(newRes.y) {}
 
     int ResolutionChangedEvent::getOldWidth() { return _oldWidth; }
